added undo function
added Undo for DB -
Converts a ResultSet to a List of Maps<String,Object> to hold the data of each row.
The String (key) represents the column name, and the Object (value) represents the value of the row.

The List represents the Table.
Does use ALOT of memory so this definitely should be improved, but unfortunately derby doesnt feature the rollback SQL query so had to come up with this frankenstein !!